requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
bandit==1.7.5
billiard==3.6.4.0
black==23.3.0
bleach==6.0.0
cachetools==5.3.1
celery==5.2.7
certifi==2023.5.7
cffi==1.15.1
cfgv==3.3.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.0
cryptography==41.0.2
defusedxml==0.7.1
distlib==0.3.6
dj-rest-auth==4.0.1
Django==4.2.1
django-allauth==0.54.0
django-auditlog==2.3.0
django-axes==6.0.5
django-celery-beat==2.5.0
django-formtools==2.4.1
django-otp==1.2.2
django-phonenumber-field==6.4.0
django-rest-auth==0.9.5
django-timezone-field==5.0
django-two-factor-auth==1.15.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-api-logger==1.1.12
drf-yasg==1.21.5
filelock==3.12.0
flake8==6.0.0
flower==1.2.0
gitdb==4.0.10
GitPython==3.1.31
humanize==4.6.0
identify==2.5.24
idna==3.4
importlib-metadata==6.6.0
inflection==0.5.1
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
kombu==5.3.0
markdown-it-py==2.2.0
MarkupSafe==2.1.3
mccabe==0.7.0
mdurl==0.1.2
mypy==1.3.0
mypy-extensions==1.0.0
nodeenv==1.8.0
oauthlib==3.2.2
packaging==23.1
pathspec==0.11.1
pbr==5.11.1
platformdirs==3.5.1
pluggy==1.0.0
pre-commit==3.3.2
prometheus-client==0.17.0
prompt-toolkit==3.0.38
psycopg2==2.9.6
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.18.0
pyflakes==3.0.1
Pygments==2.15.1
PyJWT==2.7.0
pypng==0.20220715.0
pyproject_api==1.5.1
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
qrcode==7.4.2
redis==4.5.5
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.4.1
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
six==1.16.0
smmap==5.0.0
social-auth-core==4.4.2
sqlparse==0.4.4
stevedore==5.1.0
tomli==2.0.1
tornado==6.3.2
tox==4.5.2
typeguard==4.0.0
typing_extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.2
vine==5.0.0
virtualenv==20.23.0
wcwidth==0.2.6
webencodings==0.5.1
zipp==3.15.0
